refactor: consolidate single-function files in strutil, difc, and middleware
Consolidate over-fragmented single-function files into thematic groupings:

strutil:
- Create collections.go: DeduplicateStrings + StringsToAny + CopyTrimmedStringIntMap
- Extend util.go: absorb DeepCloneJSON + InterfaceToIntString
- Consolidate corresponding test files

difc:
- Merge tags.go (StringsToTags, TagsToStrings) into labels.go
- Merge tool_classification.go (IsSingularReadTool) into evaluator.go
- Consolidate corresponding test files

middleware:
- Merge payload.go (savePayload) into jqschema.go
- Merge schema.go (inferSchema) into jqschema.go

Purely structural refactor — no logic changes.

Fixes #8297
