+++ internal/difc/evaluator.go
@@ -391,3 +391,10 @@ func (e *Evaluator) FilterCollection(
 		len(filtered.Accessible), len(filtered.Filtered), filtered.TotalCount)
 	return filtered
 }
+
+// IsSingularReadTool returns true when toolName refers to a tool expected to
+// return a single resource (e.g. get_*, *_read). List/search tools are treated
+// as collection tools even if they happen to return one item.
+func IsSingularReadTool(toolName string) bool {
+	return !strings.HasPrefix(toolName, "list_") && !strings.HasPrefix(toolName, "search_")
+}
