+++ internal/strutil/collections.go
@@ -27,3 +27,25 @@ func DeduplicateStrings(input []string, sorted bool) []string {
 	}
 	return out
 }
+
+// StringsToAny converts a []string to []interface{}.
+func StringsToAny(input []string) []interface{} {
+	out := make([]interface{}, len(input))
+	for i, value := range input {
+		out[i] = value
+	}
+	return out
+}
+
+// CopyTrimmedStringIntMap returns a defensive copy of a string→int map with
+// whitespace trimmed from all keys.
+func CopyTrimmedStringIntMap(input map[string]int) map[string]int {
+	if len(input) == 0 {
+		return nil
+	}
+	out := make(map[string]int, len(input))
+	for key, value := range input {
+		out[strings.TrimSpace(key)] = value
+	}
+	return out
+}
