--- internal/strutil/strings_to_any.go
@@ -1,10 +1,0 @@
-package strutil
-
-// StringsToAny converts a []string to []interface{}.
-func StringsToAny(input []string) []interface{} {
-	out := make([]interface{}, len(input))
-	for i, value := range input {
-		out[i] = value
-	}
-	return out
-}
